move resample function to functions
