+++ src/expensor/functions.py
@@ -10,6 +10,19 @@ import pandas as pd
 from . import constants as c
 
 
+def resample(df, period, mdate):
+    """ Resample and fill missing periods """
+
+    index = pd.date_range(df.index.min(), mdate, freq=period)
+    df = df.resample(period).sum().reindex(index).fillna(0)
+
+    # If working with years, cast the index to integer
+    if period == "YS":
+        df.index = df.index.year
+
+    return df
+
+
 def serie_to_dict(serie):
     """ Transform a serie to a dict """
